[#84] Episodes tests updates
