+++ src/tests/core/test_episodes.py
@@ -1,15 +1,16 @@
 import pytest
 from youtube_dl.utils import ExtractorError
 
+from common.enums import SourceType
 from modules.providers.exceptions import SourceFetchError
 from modules.podcast.episodes import EpisodeCreator
-from modules.podcast.models import Podcast, Episode
+from modules.podcast.models import Podcast, Episode, Cookie
 from tests.api.test_base import BaseTestAPIView
-from tests.helpers import get_podcast_data, await_
+from tests.helpers import get_podcast_data, await_, create_user
 
 
 class TestEpisodeCreator(BaseTestAPIView):
-    def test_create__ok(self, podcast, user, mocked_youtube, dbs):
+    def test_ok(self, podcast, user, mocked_youtube, dbs):
         source_id = mocked_youtube.video_id
         watch_url = f"https://www.youtube.com/watch?v={source_id}"
         episode_creator = EpisodeCreator(
@@ -23,7 +24,7 @@ class TestEpisodeCreator(BaseTestAPIView):
         assert episode.watch_url == watch_url
         assert episode.source_id == source_id
 
-    def test_create__same_episode_in_podcast__ok(self, podcast, episode, user, mocked_youtube, dbs):
+    def test_same_episode_in_podcast__ok(self, podcast, episode, user, mocked_youtube, dbs):
         episode_creator = EpisodeCreator(
             dbs,
             podcast_id=episode.podcast_id,
@@ -36,7 +37,7 @@ class TestEpisodeCreator(BaseTestAPIView):
         assert new_episode.source_id == episode.source_id
         assert new_episode.watch_url == episode.watch_url
 
-    def test_create__same_episode_in_other_podcast__ok(
+    def test_same_episode_in_other_podcast__ok(
         self, podcast, episode, user, mocked_youtube, dbs
     ):
         mocked_youtube.extract_info.return_value = {
@@ -66,7 +67,7 @@ class TestEpisodeCreator(BaseTestAPIView):
         assert new_episode.author == "Updated author"
         assert new_episode.length == 123
 
-    def test_create__same_episode__extract_failed__ok(
+    def test_same_episode__extract_failed__ok(
         self, podcast, episode, user, mocked_youtube, dbs
     ):
         mocked_youtube.extract_info.side_effect = ExtractorError("Something went wrong here")
@@ -83,7 +84,7 @@ class TestEpisodeCreator(BaseTestAPIView):
         assert new_episode.source_id == episode.source_id
         assert new_episode.watch_url == episode.watch_url
 
-    def test_create__extract_failed__fail(self, podcast, episode_data, user, mocked_youtube, dbs):
+    def test_extract_failed__fail(self, podcast, episode_data, user, mocked_youtube, dbs):
         ydl_error = ExtractorError("Something went wrong here", video_id=episode_data["source_id"])
         mocked_youtube.extract_info.side_effect = ydl_error
         episode_creator = EpisodeCreator(
@@ -96,3 +97,53 @@ class TestEpisodeCreator(BaseTestAPIView):
             await_(episode_creator.create())
 
         assert error.value.details == f"Extracting data for new Episode failed: {ydl_error}"
+
+
+class TestCreateEpisodesWithCookies(BaseTestAPIView):
+    url = "/api/podcasts/{id}/episodes/"
+    source_url = "http://link.to.source/"
+    cdata = {"data": "cookie in netscape format", "source_type": SourceType.YANDEX}
+
+    def _request(self, client, user, podcast) -> dict:
+        client.login(user)
+        url = self.url.format(id=podcast.id)
+        response = client.post(url, json={"source_url": self.source_url})
+        return self.assert_ok_response(response, status_code=201)
+
+    @staticmethod
+    def _assert_source(response_data, dbs, cookie_id):
+        episode = await_(Episode.async_get(dbs, id=response_data["id"]))
+        assert response_data["source_type"] == SourceType.YANDEX
+        assert episode.source_id == "source-id"
+        assert episode.source_type == SourceType.YANDEX
+        assert episode.cookie_id == cookie_id
+
+    def test_no_cookies_found(self, mocked_source_info, dbs, client, user, podcast):
+        # TODO: mock modules.providers.utils.get_source_media_info
+        response_data = self._request(client, user, podcast)
+        self._assert_source(response_data, dbs, cookie_id=None)
+
+    def test_specific_cookie(self, mocked_source_info, dbs, client, user, podcast):
+        cdata = self.cdata | {"created_by_id": user.id}
+        await_(Cookie.async_create(dbs, **(cdata | {"source_type": SourceType.YANDEX})))
+        cookie_yandex = await_(Cookie.async_create(dbs, **cdata))
+        response_data = self._request(client, user, podcast)
+        self._assert_source(response_data, dbs, cookie_id=cookie_yandex.id)
+        mocked_source_info.assert_called_with(self.source_url)
+
+    def test_cookie_from_another_user(self, mocked_source_info, dbs, client, user, podcast):
+        cdata = self.cdata | {"created_by_id": user.id}
+        cookie_yandex = await_(Cookie.async_create(dbs, **cdata))
+        cdata = self.cdata | {"created_by_id": create_user(dbs).id}
+        await_(Cookie.async_create(dbs, **cdata))
+
+        response_data = self._request(client, user, podcast)
+        self._assert_source(response_data, dbs, cookie_id=cookie_yandex.id)
+
+    def test_use_last_cookie(self, mocked_source_info, dbs, client, user, podcast):
+        cdata = self.cdata | {"created_by_id": user.id}
+        await_(Cookie.async_create(dbs, **cdata))
+        c2 = await_(Cookie.async_create(dbs, **cdata))
+        response_data = self._request(client, user, podcast)
+
+        self._assert_source(response_data, dbs, cookie_id=c2.id)
